dpauxmon: remove precision loss in dump_packet.
../extcap/dpauxmon.c:129:72: warning: implicit conversion loses integer
precision: 'const ssize_t' (aka 'const long') to 'guint32' (aka 'unsigned
int') [-Wshorten-64-to-32]

Reviewed-on: https://code.wireshark.org/review/27616
Petri-Dish: Dario Lombardo <lomato@gmail.com>
Tested-by: Petri Dish Buildbot

+++ extcap/dpauxmon.c
@@ -120,7 +120,7 @@ static int setup_dumpfile(const char* fifo, FILE** fp)
 	return EXIT_SUCCESS;
 }
 
-static int dump_packet(FILE* fp, const char* buf, const ssize_t buflen, guint64 ts_usecs)
+static int dump_packet(FILE* fp, const char* buf, const guint32 buflen, guint64 ts_usecs)
 {
 	guint64 bytes_written = 0;
 	int err;
@@ -328,7 +328,7 @@ static int handle_data(struct nl_cache_ops *unused _U_, struct genl_cmd *cmd _U_
 			 struct genl_info *info, void *arg _U_)
 {
 	unsigned char *data;
-	int data_size;
+	guint32 data_size;
 	guint64 ts = 0;
 	guint8 packet[21] = { 0x00 };
